requirements.txt
flask==2.2.3
Werkzeug==2.2.2
elasticsearch==7.17.9
pyyaml==6.0
python-dotenv==1.0.0
requests==2.28.2 